fix: :bug: Refactor globalSchemas import to avoid circular.
