+++ schemas/globalSchemas/articles/articleAuth.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 import { SavedArticleSchema } from "~/schemas/globalSchemas/savedArticles/savedArticle";
-import { ViewSchema } from "~/schemas/globalSchemas/views";
-import { ReactionSchema } from "~/schemas/globalSchemas/reactions";
+import { ViewSchema } from "~/schemas/globalSchemas/views/view";
+import { ReactionSchema } from "~/schemas/globalSchemas/reactions/reaction";
 
 export const ArticleAuthSchema = z.object({
   _auth: z.object({
