+++ schemas/globalSchemas/articles/indexArticle.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { ArticleFullSchema } from "~/schemas/globalSchemas/articles";
+import { ArticleFullSchema } from "~/schemas/globalSchemas/articles/article";
 import { ArticleOrderByWithRelationInputSchema } from "~/schemas/globalSchemas/articles/articleOrderBy";
 import { ArticleWhereInputSchema } from "~/schemas/globalSchemas/articles/articleWhere";
 import {
